adding tag index, fix jywarren/web#78 (#1252)
* adding tag index, fix jywarren/web#78

* migration catches and schema.rb fix

+++ db/migrate/20170120160451_drop_drupal_counter.rb
@@ -1,6 +1,8 @@
 class DropDrupalCounter < ActiveRecord::Migration
   def up
-    drop_table :node_counter
+    if table_exists? "node_counter"
+      drop_table :node_counter
+    end
   end
 
   def down
